GearUp: Integratest latest from dev.
[git-p4: depot-paths = "//depot/traktor/": change = 9068]

+++ code/Online/Impl/Tasks/TaskEnumStatistics.cpp
@@ -24,7 +24,7 @@ void TaskEnumStatistics::execute(TaskQueue* taskQueue)
 {
 	T_ASSERT (m_provider);
 	T_DEBUG(L"Online; Begin enumerating statistics");
-	std::map< std::wstring, float > statistics;
+	std::map< std::wstring, int32_t > statistics;
 	m_provider->enumerate(statistics);
 	(m_sinkObject->*m_sinkMethod)(statistics);
 	T_DEBUG(L"Online; Finished enumerating statistics");
